Move workflow in separate file.
